LZMAInputStream: Use relaxedEndCondition for .lzma files.
Now the test file good-known_size-with_eopm.lzma from XZ Utils
can be decompressed successfully.

This relies on the previous commit which ensures better
error detection and that the read position in the underlying
InputStream is correctly at the end of the .lzma file when "read"
returns -1. Without the previous commit, this commit would make
the .lzma decoder incorrectly accept these test files from XZ Utils:

    bad-too_small_size-without_eopm-1.lzma
    bad-too_small_size-without_eopm-2.lzma

This problem was fixed in XZ Utils 5.2.6 in 2022.

Thanks to the GitHub user jvanheesch for the bug report.

Fixes: https://github.com/tukaani-project/xz-java/issues/8